fix(input): add missing previous mouse position update and change double to float type of wheeldelta

+++ include/input/mouse.hpp
@@ -15,7 +15,7 @@ struct Mouse {
     bool buttonsCurrent[MOUSE_BUTTON_COUNT];
     bool buttonsPrevious[MOUSE_BUTTON_COUNT];
 
-    double wheelDelta;
+    float wheelDelta;
 
     Mouse();
